refactor: extract shared .env.saucelabs loader to reduce duplication
Move the inline env loader from both SauceLabs scripts into a shared
saucelabs-env.mjs module to satisfy SonarQube's 3% duplication limit.
