+++ scripts/saucelabs-prune-builds.mjs
@@ -19,33 +19,7 @@
  * Reads scripts/.env.saucelabs if present (same as saucelabs-app-launch-test).
  */
 
-import { existsSync, readFileSync } from 'node:fs'
-import { dirname, resolve } from 'node:path'
-import { fileURLToPath } from 'node:url'
-
-// ── Load local .env file if present ────────────────────────────────────
-
-const __dirname = dirname(fileURLToPath(import.meta.url))
-const LOCAL_ENV_PATH = resolve(__dirname, '.env.saucelabs')
-
-if (existsSync(LOCAL_ENV_PATH)) {
-  console.log(`Loading local env from ${LOCAL_ENV_PATH}\n`)
-  const lines = readFileSync(LOCAL_ENV_PATH, 'utf-8').split('\n')
-  for (const raw of lines) {
-    const line = raw.trim()
-    if (!line || line.startsWith('#')) continue
-    const eq = line.indexOf('=')
-    if (eq === -1) continue
-    const key = line.slice(0, eq).trim()
-    let val = line.slice(eq + 1).trim()
-    if ((val.startsWith("'") && val.endsWith("'")) || (val.startsWith('"') && val.endsWith('"'))) {
-      val = val.slice(1, -1)
-    }
-    if (!process.env[key]) {
-      process.env[key] = val
-    }
-  }
-}
+import './saucelabs-env.mjs'
 
 // ── Config ─────────────────────────────────────────────────────────────
